Run all GTK dialogs on one thread (#152)
Per the GDK documentation:
"GTK+, however, is not thread safe. You should only use GTK+ and GDK
from the thread gtk_init() and gtk_main() were called on. This is
usually referred to as the "main thread"."

We were previously calling GTK functions from different threads--there
was a mutex so that only one thread was calling GTK at any given time,
but that's not sufficient--only the thread *that called gtk_init* is
allowed to call GTK functions. To facilitate that, spin up a thread and
initialize GTK on it, keeping it alive for the lifetime of the program.
It's a bit unfortunate that we have to keep it around forever, but the
GTK documentation seems to indicate that once you call gtk_init on a
thread, you're stuck with that thread.