Use portals by default (#130)
* Use portals by default

It can be safely assumed that a system has portals by default; if they don't, many things will simply fail to work, including, but not limited to, the dark mode preference and many other freedesktop settings.

Those that absolutely refuse to use D-Bus will thus end up with an unusable system in the first place, if they want to use anything desktop-related. So enabling the portal backend by default is fine.

GTK will, of course, fall back to using the portals if it's running inside of a Flatpak environment, but it requires manual workarounds outside of one. And, of course, having a library depend on GTK when it's more or less unnecessary isn't ideal.

* Add default async-std feature

* Update CHANGELOG

---------

Co-authored-by: PolyMeilex <marynczakbartlomiej@gmail.com>