Support files without an extension in XDG portal filters
Due to the glob being used ("*.{file_extension}"), there was no
possibility to include files without a dot in their name when using filters.
The glob "*" is now used when an empty filter ("") or a match-all
filter ("*") is specified, which replaces globs "*." and "*.*" respectively.
The glob "*." did not make sense, as it matched files ending with a dot,
and "*.*" did not match files without an extension (which are quite popular on Linux).