fix cancel button on web backend
Before this change, clicking the "cancel" button after selecting a file
in the Web form rfd creates will still cause `pick_file().await` to
return a file successfully. I believe this is a UX bug, as a user that
clicks a "cancel" button would expect that their input will be
discarded.

So, this clears the contents of the file input element when cancelling,
which makes the later access to its `.files` property return an empty
list, making `pick_file().await` return `None` like you'd expect.